+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 HEAD (unreleased)
 -----------------
 
+- Rebranding as "Haplotype-LSO" (``hlso``).
 - Properly normalizing indels according to Tan et al. (2015).
 - Adding support for haplotyping with indels.